refactor(layered_reasoning_bot): Integrate model configuration into Bayesian strategy selection
- Updated BayesianStrategySelector to accept a ModelConfig parameter for dynamic model configuration.
- Modified the initialization of the strategy selector in LayeredReasoningBot to utilize the new model configuration.
- Adjusted model parameters in config.yaml for intent recognition and feature extraction, including temperature and max_tokens settings.
- Enhanced response generation logic to align with the updated model configuration, improving flexibility and maintainability.

These changes streamline the integration of model configurations, enhancing the adaptability of the Bayesian strategy selection process.
