+++ pocs/layered_reasoning_bot/config.yaml
@@ -11,8 +11,13 @@ model_dir: "models/"
 models:
   intent_recognition:
     model_name: "deepseek/deepseek-chat"
-    temperature: 0.7
-    max_tokens: 500
+    temperature: 0.5
+    max_tokens: 300
+  
+  feature_extraction:
+    model_name: "deepseek/deepseek-chat"
+    temperature: 0.1
+    max_tokens: 150
   
   strategy_formation:
     decision_method: "bayesian"  # Options: rule_based, bayesian, reinforcement
